refactor(app): prefer feature submodules over prefix namespaces
