+++ src/lib/features/drafts/timeline/lottery/outcome/utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
 
-import { keyForRank } from './lottery-outcome-chart-utils';
+import { keyForRank } from './utils';
 
 describe('keyForRank', () => {
   test.each([
