+++ src/lib/features/drafts/timeline/lottery/outcome/utils.ts
